refactor: 유틸 파일 분리 - utils/index를 date로 변경
