+++ front/src/components/HistoryItem/index.tsx
@@ -1,4 +1,4 @@
-import { getMonthDate } from '@utils/index';
+import { getMonthDate } from '@utils/date';
 import * as S from './styles';
 
 import ScheduleIcon from '@assets/schedule.svg';
